# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Renta de Apartamentos - API",
        "description": "Documentación de la API del sistema de rentas de apartamentos. Permite gestionar usuarios, apartamentos, quejas, contratos y más.",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "license": {
            "name": "Todos los derechos reservados - Universidad Piloto de Colombia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/edificio": {
            "get": {
                "tags": [
                    "Edificio"
                ],
                "summary": "Listar todos los edificios",
                "operationId": "f23aa2c498567bca45e53100e2a00c12",
                "responses": {
                    "200": {
                        "description": "Lista de edificios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Edificio"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Edificio"
                ],
                "summary": "Crear un nuevo edificio",
                "operationId": "fc88122c6df9ab9a3b6a7d337ee33c5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Torre Central"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle 123 #45-67"
                                    },
                                    "cantidad_pisos": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Edificio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Edificio"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/edificio/{id}": {
            "get": {
                "tags": [
                    "Edificio"
                ],
                "summary": "Obtener un edificio específico",
                "operationId": "0b0fd21e6545868fe0f36170409b0277",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edificio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edificio encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Edificio"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Edificio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Edificio"
                ],
                "summary": "Actualizar un edificio existente",
                "operationId": "aabd87839702cc5ff0ad2428bae56f9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edificio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Torre Sur"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Avenida 10 #23-45"
                                    },
                                    "cantidad_pisos": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Edificio actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Edificio"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Edificio no encontrado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Edificio"
                ],
                "summary": "Eliminar un edificio",
                "operationId": "2ec5cd34fd03127048e0356d5807f581",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edificio a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edificio eliminado correctamente"
                    },
                    "404": {
                        "description": "Edificio no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Edificio": {
                "title": "Edificio",
                "required": [
                    "nombre",
                    "direccion"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Edificio Central"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Siempre Viva 742"
                    },
                    "cantidad_pisos": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}